Python
Finished the entire project.
Imported all the modules needed for the script to
work.The script checks did the optional flag get
raised at the start,if it has it goes into the if
condition which checks does the file exists and is
it in the right format (.txt), after which it does
the conversion,if the optional flag didn't get
raised the script goes into the else condition
which gives the user an input prompt where they
can enter the cyrillic they wish to convert.